Lead_Engagement/synthetisation/requirements.txt
# Core dependencies
pydantic>=2.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0

# Data processing & cleaning
regex>=2023.10.3
python-dateutil>=2.8.2
unicode-slugify>=0.1.5

# LLM & AI
langchain>=0.1.0
unstructured
# JSON handling
ujson>=5.8.0  # Alternative plus rapide à json standard
langchain-together
google-generativeai
#rag
langchain_community
langchain-google-genai
chromadb
nltk
streamlit
